Untuk Hari ini

+++ resources/js/components/Kasir/Transaksi.vue
@@ -1,107 +1,111 @@
 <template>
     <div class="container">
-        <h1>Menu Transaksi</h1>
-        <hr>
-        <div class="row">
-            <div class="col-md-4">
-                <div class="form-group">
-                    <input type="text" class="form-control" v-model="form_transaksi_baru.no_transaksi" disabled
-                           placeholder="No Transaksi">
+        <b-breadcrumb :items="itemsTransaksi"></b-breadcrumb>
+        <b-container>
+            <h1>Menu Transaksi</h1>
+            <hr>
+            <div class="row">
+                <div class="col-md-4">
+                    <div class="form-group">
+                        <input type="text" class="form-control" v-model="form_transaksi_baru.no_transaksi" disabled
+                               placeholder="No Transaksi">
+                    </div>
+                    <div class="form-group">
+                        <input type="text" class="form-control" v-model="form_transaksi_baru.tanggal"
+                               placeholder="Tanggal"
+                               disabled>
+                    </div>
+                    <div class="form-group">
+                        <input type="text" class="form-control" v-model="form_transaksi_baru.jam" placeholder="Jam"
+                               disabled>
+                    </div>
+                    <div class="form-group">
+                        <input type="text" class="form-control" v-model="form_transaksi_baru.user" placeholder="User"
+                               disabled>
+                    </div>
                 </div>
-                <div class="form-group">
-                    <input type="text" class="form-control" v-model="form_transaksi_baru.tanggal" placeholder="Tanggal"
-                           disabled>
-                </div>
-                <div class="form-group">
-                    <input type="text" class="form-control" v-model="form_transaksi_baru.jam" placeholder="Jam"
-                           disabled>
-                </div>
-                <div class="form-group">
-                    <input type="text" class="form-control" v-model="form_transaksi_baru.user" placeholder="User"
-                           disabled>
+                <div class="col-md-3"></div>
+                <div class="col-md-5">
+                    <b-form @submit.prevent="addProdukBaru">
+                        <div class="form-group">
+                            <input type="text" placeholder="Scan Barcode atau Ketik Manual"
+                                   v-model="form_produk_tambah.barcode" class="form-control">
+                        </div>
+                    </b-form>
+                    <div class="form-group">
+                        <b-row>
+                            <b-col>
+                                <input type="text" placeholder="Nama Barang" disabled class="form-control"
+                                       v-model="form_produk_tambah.nama">
+                            </b-col>
+                            <b-col>
+                                <input type="text" placeholder="Jumlah Stock" disabled class="form-control"
+                                       :value="form_produk_tambah.jumlah_sekarang+' Buah'">
+                            </b-col>
+                        </b-row>
+                    </div>
+                    <div class="form-group">
+                        <input type="number" min="1" placeholder="Jumlah" class="form-control"
+                               v-model="form_produk_tambah.jumlah">
+                    </div>
+                    <button type="submit" class="btn btn-primary btn-sm float-right"
+                            :disabled="form_produk_tambah.barcode==null || form_produk_tambah.nama==null"
+                            @click="addProdukTambah">Tambahkan
+                    </button>
                 </div>
             </div>
-            <div class="col-md-3"></div>
-            <div class="col-md-5">
-                <b-form @submit.prevent="addProdukBaru">
+            <hr>
+            <div class="row">
+                <div class="col-md-9">
+                    <b-table
+                        responsive
+                        small
+                        :items="items"
+                        :fields="fields"
+                        striped>
+                        <template slot="harga_jual" slot-scope="row">
+                            {{ row.item.harga_jual | currency }}
+                        </template>
+                        <template slot="total" slot-scope="row">
+                            {{ row.item.harga_jual*row.item.jumlah | currency }}
+                        </template>
+                        <template slot="Option" slot-scope="row">
+                            <button class="btn btn-light btn-outline-secondary btn-sm" @click="deleteProduk(row.item)">
+                                <font-awesome-icon icon="trash"/>
+                            </button>
+                        </template>
+                    </b-table>
+                </div>
+                <div class="col-md-3">
                     <div class="form-group">
-                        <input type="text" placeholder="Scan Barcode atau Ketik Manual"
-                               v-model="form_produk_tambah.barcode" class="form-control">
+                        <label><b>Total</b></label>
+                        <input type="text" class="form-control form-control-sm" disabled :value="totalharga | currency">
                     </div>
-                </b-form>
-                <div class="form-group">
-                    <b-row>
-                        <b-col>
-                            <input type="text" placeholder="Nama Barang" disabled class="form-control"
-                                   v-model="form_produk_tambah.nama">
-                        </b-col>
-                        <b-col>
-                            <input type="text" placeholder="Jumlah Stock" disabled class="form-control"
-                                   :value="form_produk_tambah.jumlah_sekarang+' Buah'">
-                        </b-col>
-                    </b-row>
+                    <div class="form-group">
+                        <label><b>Uang</b></label>
+                        <CurrencyInput
+                            class="form-control"
+                            v-model="uang"
+                            :currency="currency_input.currency"
+                            :locale="currency_input.locale"
+                            :validate-on-input="true"
+                        />
+                    </div>
+                    <div class="form-group">
+                        <label><b>Kembalian</b></label>
+                        <input type="text"
+                               :class="'form-control form-control-lg'+(uang-totalharga<0 || totalharga==0?' is-invalid':' is-valid')"
+                               disabled :value="uang-totalharga | currency">
+                    </div>
                 </div>
-                <div class="form-group">
-                    <input type="number" min="1" placeholder="Jumlah" class="form-control"
-                           v-model="form_produk_tambah.jumlah">
-                </div>
-                <button type="submit" class="btn btn-primary btn-sm float-right"
-                        :disabled="form_produk_tambah.barcode==null || form_produk_tambah.nama==null"
-                        @click="addProdukTambah">Tambahkan
+            </div>
+            <div class="float-right">
+                <button type="submit" class=" btn btn-primary btn-sm" :disabled="uang-totalharga<0 || totalharga==0"
+                        @click="selesaiProcess">
+                    Selesai
                 </button>
             </div>
-        </div>
-        <hr>
-        <div class="row">
-            <div class="col-md-9">
-                <b-table
-                    responsive
-                    small
-                    :items="items"
-                    :fields="fields"
-                    striped>
-                    <template slot="harga_jual" slot-scope="row">
-                        {{ row.item.harga_jual | currency }}
-                    </template>
-                    <template slot="total" slot-scope="row">
-                        {{ row.item.harga_jual*row.item.jumlah | currency }}
-                    </template>
-                    <template slot="Option" slot-scope="row">
-                        <button class="btn btn-light btn-outline-secondary btn-sm" @click="deleteProduk(row.item)">
-                            <font-awesome-icon icon="trash"/>
-                        </button>
-                    </template>
-                </b-table>
-            </div>
-            <div class="col-md-3">
-                <div class="form-group">
-                    <label><b>Total</b></label>
-                    <input type="text" class="form-control form-control-sm" disabled :value="totalharga | currency">
-                </div>
-                <div class="form-group">
-                    <label><b>Uang</b></label>
-                    <CurrencyInput
-                        class="form-control"
-                        v-model="uang"
-                        :currency="currency_input.currency"
-                        :locale="currency_input.locale"
-                        :validate-on-input="true"
-                    />
-                </div>
-                <div class="form-group">
-                    <label><b>Kembalian</b></label>
-                    <input type="text"
-                           :class="'form-control form-control-lg'+(uang-totalharga<0 || totalharga==0?' is-invalid':' is-valid')"
-                           disabled :value="uang-totalharga | currency">
-                </div>
-            </div>
-        </div>
-        <div class="float-right">
-            <button type="submit" class=" btn btn-primary btn-sm" :disabled="uang-totalharga<0 || totalharga==0"
-                    @click="selesaiProcess">
-                Selesai
-            </button>
-        </div>
+        </b-container>
     </div>
 </template>
 
@@ -117,6 +121,19 @@
                     locale: 'id',
                     currency: 'IDR',
                 },
+                itemsTransaksi: [
+                    {
+                        text: 'Dashboard',
+                        to: {
+                            path: '/kasir'
+                        }
+                    },
+                    {
+                        text: 'Transaksi',
+                        active: true
+                    }
+
+                ],
                 fields: [
                     {
                         key: "code_produk",
@@ -201,6 +218,7 @@
                             this.$swal({
                                 position: 'center',
                                 type: 'error',
+                                width:300,
                                 title: 'Kode Barang Tidak Ditemukan',
                                 showConfirmButton: false,
                                 timer: 1500
@@ -213,6 +231,7 @@
                     this.$swal({
                         position: 'center',
                         type: 'error',
+                        width:300,
                         title: 'Jumlah Barang harus valid!',
                         showConfirmButton: false,
                         timer: 1500
@@ -254,13 +273,14 @@
                                 this.$swal({
                                     position: 'center',
                                     type: 'error',
+                                    width:300,
                                     title: 'Tentukan Harga Jual Produk',
                                     showConfirmButton: false,
                                     timer: 1500
                                 })
                             }
                         })
-                }else{
+                } else {
                     axios.post('api/tambahProdukTransaksi/' + this.state, {
                         id_produk: this.form_produk_tambah.id,
                         jumlah: this.form_produk_tambah.jumlah,
@@ -305,21 +325,28 @@
                     })
             },
             deleteProduk(row) {
-                axios.delete('api/deleteProduk/' + row.id+'/'+row.state, {headers: {Authorization: `Bearer ${this.$auth.getToken()}`}})
-                    .then(e => {
-                        const Toast = this.$swal.mixin({
-                            toast: true,
-                            position: 'top-end',
-                            showConfirmButton: false,
-                            timer: 3000
-                        });
-                        Toast.fire({
-                            type: 'success',
-                            title: 'Deleted'
-                        })
-                        this.fetchProduk()
-                        this.totalHarga()
-                    })
+                this.$swal.fire({
+                    title: 'Anda yakin?',
+                    text: "Menghapus Produk dari Transaksi!",
+                    type: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Ya,Hapus!'
+                }).then((result) => {
+                    if (result.value) {
+                        axios.delete('api/deleteProduk/' + row.id + '/' + row.state, {headers: {Authorization: `Bearer ${this.$auth.getToken()}`}})
+                            .then(e => {
+                                this.$swal.fire(
+                                    'Deleted!',
+                                    'Your file has been deleted.',
+                                    'success'
+                                )
+                                this.fetchProduk()
+                                this.totalHarga()
+                            })
+                    }
+                })
             },
             selesaiProcess() {
                 axios.post('api/selesaiTransaksi/' + this.form_transaksi_baru.no_transaksi, {total: this.totalharga}, {headers: {Authorization: `Bearer ${this.$auth.getToken()}`}})
